refactor: EtaSchedulingRedisTemplate service 레이어로 이동
